Refactor: consolidate AICPU register read/write into platform_regs
Merge read_reg/write_reg from separate a2a3/a2a3sim aicpu_regs.cpp
files into a unified platform_regs.cpp implementation to eliminate
code duplication and simplify maintenance.

Changes:
- Remove platform-specific aicpu_regs.{h,cpp} files
- Move read_reg/write_reg functions to platform_regs.cpp
- Enhance platform_regs.h to provide complete register management API
- Update runtime includes to use consolidated interface
