--- src/platform/a2a3sim/aicpu/aicpu_regs.cpp
@@ -1,23 +1,0 @@
-/**
- * @file aicpu_regs.cpp
- * @brief AICPU-side register access implementation (a2a3sim simulation)
- *
- * Uses volatile pointer access to host-allocated memory blocks
- * that simulate AICore hardware registers.
- */
-
-#include "aicpu/aicpu_regs.h"
-
-uint64_t read_reg(uint64_t reg_base_addr, RegId reg) {
-    volatile uint32_t* ptr = reinterpret_cast<volatile uint32_t*>(
-        reg_base_addr + reg_offset(reg));
-    __sync_synchronize();
-    return static_cast<uint64_t>(*ptr);
-}
-
-void write_reg(uint64_t reg_base_addr, RegId reg, uint64_t value) {
-    volatile uint32_t* ptr = reinterpret_cast<volatile uint32_t*>(
-        reg_base_addr + reg_offset(reg));
-    *ptr = static_cast<uint32_t>(value);
-    __sync_synchronize();
-}
